some improvement

+++ src/main/java/com/accenture/lesson/business/FbAndJoBusiness.java
@@ -3,9 +3,11 @@ package com.accenture.lesson.business;
 
 import com.accenture.lesson.response.JosephRes;
 
+import java.math.BigInteger;
+
 public interface FbAndJoBusiness {
 
-    public Integer[] showFibonacci(Integer length);
+    public BigInteger[] showFibonacci(Integer length);
 
     public JosephRes solveJosephProblem(String[] elements,Integer startIndex,Integer intervel);
 }
